backend/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
BrailleToKorean==1.0.3
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.1.1
flatbuffers==23.5.8
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
jax==0.4.9
Jinja2==3.1.2
Js2Py==0.74
keras==2.12.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyjsparser==2.7.1
pyparsing==3.0.9
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.2.0
scipy==1.10.1
selenium==4.9.1
shiboken6==6.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==2.3.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
youtube-transcript-api==0.6.0
